cfi: Add __CFI_EXCLUDE
This change introduces the __CFI_EXCLUDE macro, which can be used to
always allow indirect calls to specific functions in the core kernel
despite potential type mismatches. __CFI_EXCLUDE should be used only
for stub functions that immediately return and cannot be efficiently
called with a matching function pointer type.

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>