rust: dma_fence: Add DMA Fence abstraction
DMA fences are the internal synchronization primitive used for DMA
operations like GPU rendering, video en/decoding, etc. Add an
abstraction to allow Rust drivers to interact with this subsystem.

Note: This uses a raw spinlock living next to the fence, since we do
not interact with it other than for initialization.
TODO: Expose this to the user at some point with a safe abstraction.

Note from Deb: This patch is just for the dma-fence files which are new,
and changes that are likely to cause conflicts, mostly headers, are
separated out.

Signed-off-by: Asahi Lina <lina@asahilina.net>
Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>