rust: drm: syncobj: Add DRM Sync Object abstraction
DRM Sync Objects are a container for a DMA fence, and can be waited on
signaled, exported, and imported from userspace. Add a Rust abstraction
so Rust DRM drivers can support this functionality.

Note from Deb: This patch is just for the syncobj files which are new,
and changes that are likely to cause conflicts, mostly headers, are
separated out.

Signed-off-by: Asahi Lina <lina@asahilina.net>
Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>