workqueue: rust: add creation of workqueues
Creating workqueues is needed by various GPU drivers. Not only does it give
you better control over execution, it also allows devices to ensure that
all tasks have exited before the device is unbound (or similar) by running
the workqueue destructor.

This patch is being developed in parallel with the new Owned type [1]. The
OwnedQueue struct becomes redundant once [1] lands; at that point it can be
replaced with Owned<Queue>, and constructors can be moved to the Queue
type.

A wrapper type WqFlags is provided for workqueue flags. Since we only
provide the | operator for this wrapper type, this makes it impossible to
pass internal workqueue flags to the workqueue constructor. It has the
consequence that we need a separate constant for the no-flags case, as the
constructor does not accept a literal 0. I named this constant "BOUND" to
signify the opposite of UNBOUND.

Link: https://lore.kernel.org/r/20250325-unique-ref-v9-0-e91618c1de26@pm.me
Signed-off-by: Alice Ryhl <aliceryhl@google.com>
Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>