rust: types: Add ARef::as_inner_mut()
With DRM's GEM API, we need to be able to mutate the inner-contents of an
ARef<T> in situations where we can guarantee that there are currently no
other ARefs taken out for T. The main usecase for this at the moment is
with shmem gem objects, where there are specific fields within
kernel::drm::gem::shmem::Object<T> outside of the embedded C structure that
need to be mutated before we expose the ability to acquire ARefs for the
object.

So, introduce the unsafe as_inner_mut() function to allow for this.

Signed-off-by: Lyude Paul <lyude@redhat.com>