rust: drm: Add GPUVM abstraction
Add a GPUVM abstraction to be used by Rust GPU drivers.

GPUVM keeps track of a GPU's virtual address (VA) space and manages the
corresponding virtual mappings represented by "GPU VA" objects. It also
keeps track of the gem::Object<T> used to back the mappings through
GpuVmBo<T>.

The DriverGpuVm trait should be implemented by drivers, as they should
provide the hardware-specific implementation of the mapping/unmapping
and remapping steps. The steps themselves are computed by the GPUVM C
code.

This abstraction for drivers looking to support bind and unbind
operations (known as the VM_BIND API) commonly used to implement Vulkan
'sparse resources', which includes all upcoming Rust GPU drivers so far.

We require &mut self for any operations that modify the interval tree
used to keep track of the current bindings in a VM. In this sense, Rust
drivers are expected to provide some kind of interior mutability in
order to use any operation that would otherwise race to access the tree.

Locking the BO va list is still todo, although it should be
implementable in a future iteration of this patch, as soon as the
pre-requisite patches land. Asynchronous VM_BINDS should also be
possible.

This patch notably misses any other type of locking helpers, like any
drm_exec or prepare_vm helpers. That's because it is not clear how these
should look like so far, such that it is safer to proceed with the
current subset of operations, which are deemed both somewhat sufficient
and required for upcoming drivers.

Note that all operations provided were tested on a real driver and on a
real device. Specifically, all operations so far are required.

Signed-off-by: Asahi Lina <lina@asahilina.net>
Co-developed-by: Daniel Almeida <daniel.almeida@collabora.com>
Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>