rust: gpuvm: add call to prepare
This call ensures that all objects within the VM have been locked. This is
needed during submits.

Note that this commit is not aimed at upstream, at least not for now. This
is because  this part of the GPUVM feature set will probably be handled
either by the new scheduler (being worked on by RedHat) or by a subsequent
patch once it becomes clear how it should be abstracted.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>