tyr: fix broken map/unmap algorithms
These had an array of issues:

- the results of (map/unmap)_pages was never checked
- we did not operate under the assumption that (map/unmap)_pages may not
  grant the entire request at once
- only 4k pages were supported

In practice this meant that ranges would sometimes not be mapped or
unmapped, even though we would print a trace saying they were, which
would then be _very_ confusing when iommu page faults happened.

Just copy the thing in Panthor 1:1 and be done with this.

In particular, we now have a _loop_ in place to ensure that we call
(map/unmap)_pages for as long as needed while properly updating both the
physical address and the corresponding iova.