tyr: set maximum opp
We don't have any power management code at the moment, so a sensible
default is to choose the maximum clock/voltage configuration.

The device will be at its maximum performance point at all times (even
when the workload doesn't require it) but at least we can use it for
testing and benchmarks.