tyr: don't mark groups as idle for now
This actually does not work, since it doesn't properly handle
CS_EXTRACT_INIT, which makes the GPU re-execute the ringbuf from the
start whenever we reschedule the group and make it active.

Additionally, we shouldn't really evict idle groups if there's nothing
to schedule in their place anyways.

Disable this behavior while we don't properly fix it.