tyr: remove broken poll condition in ring_csg_doorbell
Not only is the condition wrong - we would have to wait until the ack
flipped, but we had hardcoded ack==1 instead - but to top it off, we
don't need to wait at all here.

Remove this to get rid of the massive 100ms delay before completing
fences. This lets us go back to 30fps in weston-simple-egl and almost
1000fps in kmscube at 4k.

This fix apparently also makes the slowdown that we got after repeatedly
running gfx apps disappear.