Tyr: store TyrData in Arc
When TyrDriver was introduced in:
`cf4fd52e3236 rust: drm: Introduce the Tyr driver for Arm Mali GPUs`

TyrDriver implemented the drm::Driver associated type Data as a plain
TyrData struct

  impl drm::Driver for TyrDriver {
      type Data = TyrData;

But enqueuing work require the trait `RawWorkItem<ID> to be implemented

  pub fn enqueue<W, const ID: u64>(&self, w: W) -> W::EnqueueOutput
  where
      W: RawWorkItem<ID> + Send + 'static,

The trait `RawWorkItem<ID> is implemented for Arc

    unsafe impl<T, const ID: u64> RawDelayedWorkItem<ID> for Arc<T>
    where
        T: WorkItem<ID, Pointer = Self>,
        T: HasDelayedWork<T, ID>,

So the drm::Driver associated Data type, TyrData, needs to be stored in
an Arc. In fact, this was in the tyr-next branch all along, I just
missed it when reconciling with upstream.