tyr: wrap TyrData in Ok() when creating drm::Device
When creating a new drm::Device, the second argument has to implement
`PinInit<T::Data, Error>`

    pub fn new(dev: &device::Device, data: impl PinInit<T::Data, Error>) -> Result<ARef<Self>> {

Since TyrData is stored in an Arc when it is the associated Data type
for the drm::device, and since Arc<TyrData> doesn't implement PinInit,
there is a compile error if we try to pass just Arc<TyrData> to
drm::Device::new()

Since `Result<T, E>` implements `PinInit<T, E>`, wrapping the Arc in `Ok(...)`
allows the data to be written into the pinned DRM device correctly.

This was in the tyr-next branch but I missed it when reconciling with
upstream.