tyr: don't signal fences directly on timeout
If other jobs are submitted, we will just signal the fence in the
CSG_SYNC_UPDATE handler.

This is incorrect, as the right thing to do would be to reset the GPU and
then spawn a work to process all leftover fences, setting an error on all
of them.

Still, this is good enough for a demo.